requirements.txt
paho-mqtt
pybluez
toml
pexpect
